2.0.70
Marking nodes visited should work better
DungeonExplorer MiniMap logic will now always navigate directly to the "Objective" first (if visible)
Fixed another stuck with MoveToMapMarker in rifts
Added <MoveToObjective questId="1" /> tag - will move to the bounty/rift/quest objective if visibile on the minimap, and interact with the portal (if it's a portal. Thanks XZJV for the code and idea!